chore: remove all next/router router
