+++ packages/toolkit/src/components/card-pipeline/Footer.tsx
@@ -102,7 +102,6 @@ export const Footer = ({
               </Button>
             }
             pipeline={pipeline}
-            router={router}
           />
         ) : (
           <Button
