+++ packages/toolkit/src/components/top-bar/TopbarDropdown.tsx
@@ -1,0 +1,32 @@
+"use client";
+
+import * as React from "react";
+import { DropdownMenu } from "@instill-ai/design-system";
+
+export const TopbarDropdownItem = ({
+  children,
+  asChild,
+  onClick,
+}: {
+  children: React.ReactNode;
+  asChild?: boolean;
+  onClick?: () => void;
+}) => {
+  return (
+    <DropdownMenu.Item
+      asChild={asChild}
+      className="cursor-pointer !gap-x-2 !px-2.5 !py-[9px] !product-body-text-3-medium hover:!bg-semantic-bg-base-bg hover:!text-black"
+      onClick={onClick}
+    >
+      {children}
+    </DropdownMenu.Item>
+  );
+};
+
+export const TopbarDropdownGroup = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  return <DropdownMenu.Group className="!p-1.5">{children}</DropdownMenu.Group>;
+};
