+++ packages/toolkit/src/components/top-bar/index.ts
@@ -1,2 +1,1 @@
-export * from "./Topbar";
 export * from "./AppTopbar";
