+++ packages/toolkit/src/lib/hook/index.ts
@@ -1,6 +1,5 @@
 export * from "./useGuardPipelineBuilderUnsavedChangesNavigation";
 export * from "./useCreateUpdateDeleteResourceGuard";
 export * from "./useRefSize";
-export * from "./useNavigationObserver";
 export * from "./useWindowSize";
 export * from "./useOnScreen";
