+++ packages/toolkit/src/lib/type/general.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import * as React from "react";
-import { NextRouter } from "next/router";
 import { UseFormReturn } from "react-hook-form";
 import { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
 
@@ -27,12 +26,6 @@ export type GeneralUseFormReturn = UseFormReturn<
 export type GeneralRecord = Record<string, any>;
 
 export type InstillAppEnv = "APP_ENV_CORE" | "APP_ENV_CLOUD";
-
-export type GeneralPageProp = {
-  router: NextRouter;
-  enableQuery: boolean;
-  accessToken: Nullable<string>;
-};
 
 export type GeneralAppPageProp = {
   router: AppRouterInstance;
