+++ packages/toolkit/src/view/pipeline/view-pipeline/Head.tsx
@@ -334,7 +334,6 @@ export const Head = ({
                       </Button>
                     }
                     pipeline={pipeline.data}
-                    router={router}
                   />
                 ) : (
                   <Button
